f5-requirements.txt
accelerate>=0.33.0
bitsandbytes>0.37.0; platform_machine != 'arm64' and platform_system != 'Darwin'
cached_path
click
datasets
ema_pytorch>=0.5.2
hydra-core>=1.3.0
jieba
librosa
matplotlib
numpy<=1.26.4
pydub
pypinyin
safetensors
soundfile
tomli
torch<2.7.0
torchaudio<2.7.0
torchdiffeq
tqdm>=4.65.0
transformers
transformers_stream_generator
vocos
wandb
x_transformers>=1.31.14
fastapi
uvicorn 
numpy 
requests
slowapi
pydantic_settings